Cython implementations for to_allele_counts (#288)
* cython implementations for to_allele_counts

* use optimisations for to_allele_counts

* regenerated model.c with to_allele_counts functions

* recythonize

* release note

* Use uint8_t for to_allele_counts and remove dtype argument.

* Remove use of dtype argument when calling to_allele_counts

* re-cythonize allel/opt/model.pyx

+++ allel/stats/mendel.py
@@ -187,8 +187,8 @@ def mendel_errors(parent_genotypes, progeny_genotypes):
 
     # transform into per-call allele counts
     max_allele = max(parent_genotypes.max(), progeny_genotypes.max())
-    parent_gc = parent_genotypes.to_allele_counts(max_allele=max_allele, dtype='i1')
-    progeny_gc = progeny_genotypes.to_allele_counts(max_allele=max_allele, dtype='i1')
+    parent_gc = parent_genotypes.to_allele_counts(max_allele=max_allele).astype('i1')
+    progeny_gc = progeny_genotypes.to_allele_counts(max_allele=max_allele).astype('i1')
 
     # detect nonparental and hemiparental inheritance by comparing allele
     # counts between parents and progeny
